Normative: Re-add element finishers
Finishers that come from elements can be useful for letting field
and method declarations attach runtime metadata to classes. This
proposal had removed the previous proposal's element finishers,
but that was an attempt at minimalism without taking this use
case into account.

This PR also changes the order of properties, putting the extras
all at the end, after the properties that were declared in the
class syntactically.

Closes #20