Normative: Check for overlapping elements from decorators
With this patch, if decorators result in multiple class elements with
the same name, a TypeError is thrown during execution of the class
definition. The purpose of the change is to avoid programming errors:
if decorators create overlapping class elements, it would be possible
to implement a "last element wins" policy, similar to how class literals
look. However, it's very easy to accidentally implement multiple
decorators that use the same fixed field name for internal storage; in
this case, the decorators may work independently but break in strange
ways when used together. Programmers are recommended to use a class
decorator instead if a single key is needed for internal storage for
multiple decorators.

Note that multiple syntactic definitions of the same class (public)
class element are still permitted, including in decorated classes. The
patch only bans decorators synthesizing overlapping class elements.

Overlapping class element errors are reported to the programmer
immediately when an overlap is created by a particular decorator,
rather than after all decorators are run. The hope is that this will
assist in debugging, e.g., by ensuring that further decorators will
be continuing to operate on an invalid element list, and making it
easier for implementations to provide

Addresses #8

NOTE: this should sit on top of a coalescing patch, which I thought
I already wrote.