Normative: Early error for multiply decorated method pairs
Getter/setter pairs are decorated by decorating one of the two; the
decorator will be applied to the pair together after coalescing.
Decorating both creates a runtime error in #61--in general, you
can't do better than a runtime error, due to computed property names.
This patch adds early errors for the "easy cases" where it's apparent
that there is this collision ahead of time--where the property name
is a literal or private.

Fixes #60