Normative: Use a non-frozen object to represent Private Names
This patch changes PrivateName from a primitive type to a non-frozen
object. Rather than passing the `get` and `set` functions to decorators,
the PrivateName constructor is instead passed, and is no longer present
as a property of the global object.

The change is hoped to reduce implementation complexity compared to the
previous primitive type semantics, and describes a possibility which
was discussed at the March 2018 TC39 meeting.

Applications which need an unmodified version of PrivateName's methods
are recommended to make a copy of them early on before anything modifies
them, similarly to any other JS built-in.

Addresses #68