Normative: Use own properties for privateName objects
This patch changes PrivateName objects to have a null prototype
and own methods, as described in

https://github.com/tc39/proposal-decorators/issues/68#issuecomment-403655010

I'm not sure whether we want to do the change, but I'm writing it up
so that we can think more carefully about whether we want to go in
this direction. This patch is the most concrete, reasonable thing
I can imagine in the direction of more integrity. It's also a relatively
small change vs the previous proposal.

Some results of the change:
- It's not possible to effectively monkey-patch anything in particular
  to intercept private name accesses; you can only change one in
  particular which you already have access to.
- There may be slightly more memory consuption due to the privateName
  objects having more own properties, but not that much, as the methods
  remain not bound to the receiver.

Closes #68