Normative: Remove PrivateName constructor from decorators
The constructor can be polyfilled in 5 lines of code; there may
be a cleaner way to include it in the standard library in the
future, or it may be made unnecessary in favor of certain future
syntax. Therefore, it is not necessary to pass it as the second
argument for all decorators.

This patch also updates the code samples in the documentation
(some of which was a bit out of date).