Normative: Remove class decorator access to private names
This patch makes it so that class decorators do not see private
class elements. With this patch, only element decorators have
the ability to do metaprogramming on private class elements.

For field/method decorators, granting access to the private name
seems syntactically clear and implied by the construct. There
are also several critical use cases, including:
- Limited-capability accessors (@reader)
- Observation pattern (@tracked)
- Friends/testable (@testable)

For access from class decorators to private, different people
have had different intuitions: Some expect access, while others
say it violates privacy. A workaround here, under this patch,
is to use an element decorator for each private element, rather
than a class decorator, to accomplish the same thing. This
may be ugly, but it works for all the use cases I'm aware of.

As a possible follow-on proposal, a specifically "trusted"
type of decorator could be created, though the details are unclear.

See #133