Normative: Remove errant ToPrimitive
Looks like this ToPrimitive call is a relic from back when PrivateName
was a primitive type. Now that it's an object, and toString() throws,
the result would be that an exception would be thrown on this line.
Furthermore, it's redundant given the following ToPropertyKey call.
This patch simply removes the ToPrimitive operation.