DO NOT MERGE: Remove elements in favor of class extras
Initially, the decorators proposal had class descriptors which
contained an elements array with public and private elements.
Later, concerns were raised about exposing this private information
unexpectedly (#133), and this access was removed (#151).

However, this patch caused a problem: private field initializers
were accidentally sorted after public field initializers, an observable
and counterintuitive change (#183). Halfway solutions (#184) did
not really work out; we added back the full element support
in https://github.com/tc39/proposal-decorators/issues/133#issuecomment-452676378 .

Although we believe that it makes sense for decorators to be able
to see into class bodies, including private, and there are many
important use cases (for both public and private, #192), this support
doesn't have strict consensus in committee, so this patch serves
as a conservative option in case including elements is unacceptable.

This is not the preferred alternative, but it's a well-articulated
backup plan. I'm writing this out in advance in the interest of time,
to ensure that we have a path to Stage 3 and that this does not block
things.