Normative: Disallow repeat use of private names
This patch restricts private names to only be used in a class
definition "once"--that is, if a decorator gets ahold of a PrivateName
and tries to use it for multiple different class elements (in the
same class or in different classes), a TypeError will be thrown when
the class is being defined.

Rationale:
- Private names should be easier to think about if they always mean
  the same thing.
- Many proposed use cases for multiple use of the same private name
  would leak the name more than intended.
- This restriction supports certain implementation strategies for
  private methods based on "brand checking" rather than actually
  including the method in the instance.

Includes rebase on top of current private methods spec, which meant
copying (and slightly modifying) a bunch of spec text from a previous
revision of private methods.