Normative: replacers -> static fields -> finishers
If a replace hook outputs a new class, it would be surprising if
static field initializers saw the old, underlying class. This patch
runs replace hooks before static field initializers, in order to
provide the expected behavior. Finish hooks still run afterwards,
so that they have a view on the entire class. As part of this change,
a single hook can have both finish and replace callbacks.

Closes #211