Update Metadata access API and examples
Updates the way that metadata is accessed to use a new API,
`Object.getMetadata`, rather than `Symbol.metadata`. This API makes
metadata less open-by-default, as users won't be able to discover
metadata by iterating over the known keys. The key has also been updated
to be _required_ to be a symbol. This helps to prevent collisions with
alternative libraries/keys, and also helps to ensure that metadata is
not open-by-default as the key must be exported in order to access
metadata defined with it.

In addition, there are a few more changes:

1. The format that metadata is stored in has been updated to be based
   around, such that when you load metadata with
   `Object.getMetadata(target, key)`, you get _all_ of the metadata for
   that target defined with `key`, and _only_ the metadata defined with
   `key`. This allows authors to quickly query their own metadata,
   without iterating through properties that have other metadata defined
   on them.

2. Private metadata has been updated to be an array rather than an
   object, since the name/spelling of private elements is not
   consequential for the usage of their metadata.

3. Metadata objects have been updated to inherit from their parent
   metadata, making it easier to read and iterate all metadata in the
   class for a given key.

4. Decorators for private elements now also recieve a `name` property,
   which is the spelling of the private element. This is solely for
   debugging purposes and cannot be used in any meaningful way
   otherwise.

5. Private field access example has been updated to a more relevant
   example, dependency injection on private class fields.