Remove `init` kinds and add `addInitializer`
Removes the `init` kinds in favor of an `addInitializer` method which is
optionally available on the context object passed to decorators. This
method is only present if the `@init:` modifier was used on the
decorator, and can be used to add an initializer for the element which
runs after any other pre-existing initializers.

The name choice here is a bit confusing, as it overlaps with class field
and auto-accessor initializers (e.g. the right hand side of a class
field). However, alternatives such as `finisher` do not really tie into
the `@init:` modifier's name, and `@finish:` or alternatives seem worse
from a DX perspective. I don't have a very strong opinion here, but I
do think it would be better for these to line up.