Remove Metadata and access
At the plenary today consensus was reached to move this proposal to
stage 3 conditionally, with metadata being split out into a separate
stage 2 proposal. The notion of "access" and the `access` object are
highly related to metadata (it's pretty rare that you would have a use
case for access outside of metadata), so I'm also removing it in this PR
and will be adding it back in the stage 2 metadata proposal.